+++ CHANGELOG.md
@@ -10,6 +10,7 @@ dd mm yyyy
 - Services: Service accounts are now reloaded after a manual update (frontend).
 - New: Charts with statistics on player logins and app requests.
 - ESI: Added check for "Undefined 429 response", see also https://github.com/esi/esi-issues/issues/1227
+- ESI: Do not query different types at once via "universe/names".
 - Tracking: Trim search.
 - Tools: Added shell script to build the distribution with docker.